src\tools\troubleshoot_db.py  simplifiable-if-expression
if result_proxy.fetchall() == [(123,)] can result in only True or False, the external condition is not needed.
However, if  result_proxy.fetchall() might return Noe, the external condition can translate it to False.

The external condition hurt readability since it takes a bit to verify what it does.
bool is a simpler implementation

See
https://stackoverflow.com/questions/76094401/the-if-expression-can-be-replaced-with-test-simplifiable-if-expression